+++ tests/unit/connector/test_http_output.py
@@ -1,10 +1,11 @@
 # pylint: disable=missing-docstring
 import json
-from unittest import mock
+from unittest.mock import MagicMock, patch
 
 import pytest
 import requests
 import responses
+from requests.exceptions import SSLError
 
 from tests.unit.connector.base import BaseOutputTestCase
 
@@ -18,6 +19,7 @@ class TestOutput(BaseOutputTestCase):
         "target_url": TARGET_URL,
         "user": "user",
         "password": "password",
+        "verify": "False",
     }
 
     expected_metrics = [
@@ -44,15 +46,15 @@ class TestOutput(BaseOutputTestCase):
         self.object.metrics.number_of_processed_events = 0
         self.object.metrics.number_of_http_requests = 0
         responses.add(responses.POST, f"{TARGET_URL}/123", status=404)
-        events = [{"event1_key": "event1_value"}, {"event2_key": "event2_value"}]
-        batch = ("/123", events)
+        events = '{"event1_key": "event1_value", "event2_key": "event2_value"}'
+        batch = f"/123,{events}"
         self.object.store(batch)
-        assert self.object.metrics.number_of_failed_events == 2
+        assert self.object.metrics.number_of_failed_events == 1
         assert self.object.metrics.number_of_processed_events == 0
         assert self.object.metrics.number_of_http_requests == 1
 
     @pytest.mark.parametrize(
-        "testcase, input_data, number_of_expected_requests, number_of_expeted_events",
+        "testcase, input_data, number_of_expected_requests, number_of_expected_events",
         [
             ("dict to root", {"message": "my event message"}, 1, 1),
             ("dict to target", ("/abc", {"message": "my event message"}), 1, 1),
@@ -72,7 +74,7 @@ class TestOutput(BaseOutputTestCase):
     )
     @responses.activate
     def test_store_counts_number_of_requests_and_events(
-        self, testcase, input_data, number_of_expected_requests, number_of_expeted_events
+        self, testcase, input_data, number_of_expected_requests, number_of_expected_events
     ):
         if isinstance(input_data, tuple):
             target_url = input_data[0]
@@ -87,7 +89,7 @@ class TestOutput(BaseOutputTestCase):
         assert (
             self.object.metrics.number_of_failed_events == 0
         ), f"no failed events for input {testcase}"
-        assert self.object.metrics.number_of_processed_events == number_of_expeted_events
+        assert self.object.metrics.number_of_processed_events == number_of_expected_events
         assert self.object.metrics.number_of_http_requests == number_of_expected_requests
 
     @pytest.mark.parametrize(
@@ -141,9 +143,35 @@ class TestOutput(BaseOutputTestCase):
     def test_store_counts_processed_events(self):
         responses.add(responses.POST, f"{TARGET_URL}/")
         self.object.metrics.number_of_processed_events = 0
-        self.object.store({"message": "my event message"})
+        self.object.store("," + '"message": "my event message"')
         assert self.object.metrics.number_of_processed_events == 1
 
     @pytest.mark.skip(reason="not implemented")
     def test_setup_calls_wait_for_health(self):
         pass
+
+    @responses.activate
+    def test_respond_verify_unverified_host(self):
+        with patch("requests.post", side_effect=SSLError("SSL certificate verify failed")):
+            with patch.object(self.object.__class__, "verify", new=True):
+                try:
+                    self.object.store("," + '"message": "my event message"')
+                except SSLError as e:
+                    assert "SSL certificate verify failed" in str(e)
+
+    @responses.activate
+    def test_response_verify_verified_host(self):
+
+        mock_response = MagicMock()
+        mock_response.status_code = 200
+        mock_response.text = "Success"
+
+        with patch("requests.post", return_value=mock_response) as mock_post:
+            with patch.object(self.object.__class__, "verify", new=True):
+                self.object.store("," + '"message": "my event message"')
+
+                mock_post.assert_called_once()
+                assert mock_post.call_args.kwargs["verify"] is True
+
+                assert mock_response.status_code == 200
+                assert mock_response.text == "Success"